Move convertToBiblatex back to CleanupAction as it has GUI dependencies
