+++ src/main/java/net/sf/jabref/gui/actions/CleanUpAction.java
@@ -18,6 +18,7 @@ package net.sf.jabref.gui.actions;
 import java.io.File;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 
 import javax.swing.JCheckBox;
@@ -118,6 +119,40 @@ public class CleanUpAction extends AbstractWorker {
         String origValue = entry.getField(fieldName);
         compound.addEdit(new UndoableFieldChange(entry, fieldName, origValue, ""));
         entry.setField(fieldName, "");
+    }
+
+    /**
+     * Converts to BibLatex format
+     */
+    public static void convertToBiblatex(BibtexEntry entry, NamedCompound ce) {
+        for (Map.Entry<String, String> alias : EntryConverter.FIELD_ALIASES_TEX_TO_LTX.entrySet()) {
+            String oldFieldName = alias.getKey();
+            String newFieldName = alias.getValue();
+            String oldValue = entry.getField(oldFieldName);
+            String newValue = entry.getField(newFieldName);
+            if ((oldValue != null) && (!oldValue.isEmpty()) && (newValue == null)) {
+                // There is content in the old field and no value in the new, so just copy
+                entry.setField(newFieldName, oldValue);
+                ce.addEdit(new UndoableFieldChange(entry, newFieldName, null, oldValue));
+
+                entry.setField(oldFieldName, null);
+                ce.addEdit(new UndoableFieldChange(entry, oldFieldName, oldValue, null));
+            }
+        }
+
+        // Dates: create date out of year and month, save it and delete old fields
+        if ((entry.getField("date") == null) || (entry.getField("date").isEmpty())) {
+            String newDate = entry.getFieldOrAlias("date");
+            String oldYear = entry.getField("year");
+            String oldMonth = entry.getField("month");
+            entry.setField("date", newDate);
+            entry.setField("year", null);
+            entry.setField("month", null);
+
+            ce.addEdit(new UndoableFieldChange(entry, "date", null, newDate));
+            ce.addEdit(new UndoableFieldChange(entry, "year", oldYear, null));
+            ce.addEdit(new UndoableFieldChange(entry, "month", oldMonth, null));
+        }
     }
 
     private void initOptionsPanel() {
@@ -319,7 +354,7 @@ public class CleanUpAction extends AbstractWorker {
                 doConvertUnicode(entry, ce);
             }
             if (choiceConvertToBiblatex) {
-                EntryConverter.convertToBiblatex(entry, ce);
+                convertToBiblatex(entry, ce);
             }
 
             ce.end();
